requirements.txt
agentops==0.3.12
crewai==0.63.6
python-dotenv==1.0.1